Fix IMU report
The order of acceleration and angular velocity fields are fixed so the
reported values are correct. The axis orientation is positive X pointing
right, positive Y pointing up, and positive Z pointing towards the user.

For the angular velocities, the following convention is used for the
names (following the right thumb rule):

Roll: rotation around X axis
Pitch: rotation around Y axis
Yaw: rotation around Z axis

The code has been tested by checking that the sign of the values are
consistent with the axis orientation when translating/rotating the
device.

References:
- https://gamedev.stackexchange.com/a/87178
- http://eleccelerator.com/wiki/index.php?title=DualShock_4#Report_Structure